+++ cfdm/test/test_read_write.py
@@ -1256,6 +1256,13 @@ class read_writeTest(unittest.TestCase):
             f = cfdm.read("test_read_write.py", file_type=file_type)
             self.assertEqual(len(f), 0)
 
+    def test_write_chunk_cache(self):
+        """Test the cfdm.write 'chunk_cache' keyword."""
+        f = self.f0
+        for fmt in ("NETCDF3_CLASSIC", "NETCDF4"):
+            for chunk_cache in (None, 4194304):
+                cfdm.write(f, tmpfile, fmt=fmt, chunk_cache=chunk_cache)
+
 
 if __name__ == "__main__":
     print("Run date:", datetime.datetime.now())